+++ main.py
@@ -11,6 +11,7 @@ from modules.gameloop import where
 from modules.platforms import win
 from modules.reconnects import game_closed
 from modules.resolution import gen_images_new_resolution
+from modules.utils import rsleep
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
@@ -45,13 +46,13 @@ def main():
                 else:
                     BNCount = 0
             else:
-                time.sleep(1)
+                rsleep(1)
         except KeyboardInterrupt as kerr:
             log.info("Keyboard Interrupt %s", kerr)
             sys.exit(0)
         except Exception as error:
             log.exception("Error: %s", error)
-            time.sleep(1)
+            rsleep(1)
 
 
 if __name__ == "__main__":